+++ src/main/java/domain/coordinate/position/Row.java
@@ -1,12 +1,12 @@
-package domain.position;
+package domain.coordinate.position;
 
 import java.util.Objects;
 
-public class Column {
+public class Row {
 
     private final Position position;
 
-    public Column(Position position) {
+    public Row(Position position) {
         this.position = position;
     }
 
@@ -14,12 +14,16 @@ public class Column {
         position.moveBy(distance);
     }
 
-    public int getColumnDifference(Column column) {
-        return position.getMinusPosition(column.position);
+    public int getRowDifference(Row row) {
+        return position.getMinusPosition(row.position);
     }
 
-    public Column copied() {
-        return new Column(position.copied());
+    public Row copied() {
+        return new Row(position.copied());
+    }
+
+    public boolean isSamePosition(int otherPosition) {
+        return position.isSame(otherPosition);
     }
 
     @Override
@@ -30,8 +34,8 @@ public class Column {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Column column = (Column) o;
-        return Objects.equals(position, column.position);
+        Row row = (Row) o;
+        return Objects.equals(position, row.position);
     }
 
     @Override
