+++ src/test/java/domain/coordinate/position/PositionTest.java
@@ -1,4 +1,4 @@
-package domain.position;
+package domain.coordinate.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
